requirements.txt
matplotlib==3.1.2
numpy==1.17.4
openslide_python==1.4.1
scipy==1.14.1
skimage==0.0
